ci: auto-apply Supabase migrations on merge to main
- New workflow .github/workflows/deploy-migrations.yml: path-filtered
  on supabase/migrations/**, runs supabase db push --linked, with a
  workflow_dispatch dry-run option and a concurrency group that queues
  rather than cancels (canceling a half-applied migration is unsafe).
- docs/deploy-migrations.md mirrors deploy-edge-functions.md so the
  two pipelines feel consistent; CLAUDE.md doc map points at it.
- scripts/pre-deploy-check.sh gains an advisory warning for
  uncommitted files under supabase/migrations/ — the workflow applies
  every committed file on the next push, so visibility before pushing
  matters.
- supabase/migrations/20260424234740_rewrite_get_subscriber_feed.sql
  gains explicit ::integer casts on a.story_magnitude at both
  calculate_relevance call sites, matching the version applied to
  prod via MCP. Behavior identical (Postgres widens smallint→integer
  implicitly), but keeps repo and prod byte-aligned so a future
  re-apply doesn't silently revert prod to the implicit form.

NOTE: requires `supabase migration repair --status applied
20260424234737..20260424234740` to be run locally against prod
BEFORE merge — the four PR #15 migrations were applied via MCP and
the schema_migrations tracker doesn't yet know about them. See PR
description and docs/deploy-migrations.md.

+++ supabase/migrations/20260424234740_rewrite_get_subscriber_feed.sql
@@ -90,7 +90,7 @@ BEGIN
               a.published_at, a.consensus_signal, a.extracted_tickers,
               a.inference_watch,
               calculate_relevance(
-                a.story_magnitude,
+                a.story_magnitude::integer,
                 -- Market: max() across primary_entities. Missing ticker
                 -- defaults to 0.2 (small-cap). Macro entities get 0.9.
                 COALESCE((
@@ -192,7 +192,7 @@ BEGIN
               a.published_at, a.consensus_signal, a.extracted_tickers,
               a.inference_watch,
               calculate_relevance(
-                a.story_magnitude,
+                a.story_magnitude::integer,
                 COALESCE((
                   SELECT MAX(
                     CASE
